15_A.I. Based Document Digitization/Source-Code/Group - 15 - A.I. Based Document Digitization Backend/requirements.txt
blis==0.4.1
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
cymem==2.0.3
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-vision==1.0.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.9
murmurhash==1.0.2
numpy==1.20.2
plac==1.1.3
preshed==3.0.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
spacy==2.2.3
srsly==1.0.2
thinc==7.3.1
tqdm==4.43.0
urllib3==1.25.8
wasabi==0.6.0